Add files via upload
Within the testing apparatus, there are two rods used for mounting the object to be tested. On each of these rods, there is one set of strain gauges in full bridge configuration. Full bridge configuration is used as it has maximum sensitivity and is temperature compensated. The differential output of the bridge is connected to an instrumentation amplifier for amplification and noise reduction. Note that the output of this entire circuit yields an analog input between 0-5 V.